# perf: avoid sstable info copy in event handler