@hzxa21 2024-03-20
Test on a single-CN cluster with 2w+ actors created.
- `on_sync_finish` event handler latency p50 decreases from 1.67s to 307ms
- `barrier_latency` p99 decreases from 7.38s to 1.67s.

Main:
![image](https://github.com/risingwavelabs/risingwave/assets/5518566/a9675401-2d0b-48f2-ad61-f19cfd998e9e)
![image](https://github.com/risingwavelabs/risingwave/assets/5518566/2b35880b-d008-4f35-bdaf-0e284cb7d4fa)

PR:
![image](https://github.com/risingwavelabs/risingwave/assets/5518566/945e69bb-d0d7-41df-8499-e4aa61d7c44e)
![image](https://github.com/risingwavelabs/risingwave/assets/5518566/e488a29d-7d73-4b8e-bcc8-633012b6c636)